ROC-guided Regression Forest

Fits a "rocForest" model.

rocForest(formula, data, id, subset, splitBy = c("CON", "dCON"),
  control = list())

Arguments

formula
a formula object, with the response on the left of a '~' operator, and the terms on the right. The response must be a survival object returned by the 'Surv' function.

data
an optional data frame in which to interpret the variables occurring in the 'formula'.

id
an optional vector used to identify time dependent covariate. If missing, then each individual row of 'data' is presumed to represent a distinct subject and each covariate is treated as a baseline covariate. The length of 'id' should be the same as the number of observations.

subset
an optional vector specifying a subset of observations to be used in the fitting process.

splitBy
a character string specifying the splitting algorithm. See *Details*.

control
a list of control parameters. See 'details' for important special features of control parameters.

Value

An object of S3 class "rocForest" representing the fit, with the following components:

Examples

library(survival)
set.seed(1)
dat <- simu(100, 0, 1.3)
fit <- rocForest(Surv(Time, death) ~ z1 + z2, id = id, data = dat,
        control = list(minsp = 3, minsp2 = 1, B = 50))
fit

#> rocForest result
#> 
#> Call:
#>  rocForest(formula = Surv(Time, death) ~ z1 + z2, data = dat,      id = id, control = list(minsp = 3, minsp2 = 1, B = 50)) 
#> 
#> Sample size:                                  100 
#> Number of independent variables:              2 
#> Number of trees:                              50 
#> Split rule:                                   CON 
#> Number of variables tried at each split:      2 
#> Minimum number of failure in a node:          3 
#> Minimum number of failure in a teminal node:  1 

## Print individual trees
print(fit, 1)

#>  ROC-guided survival forest: tree # 1 
#> 
#>  node), split
#>    * denotes terminal node
#>   
#> Root                              
#>  ¦--2) z1 <= 0.27473              
#>  ¦   ¦--4) z1 <= 0.14286          
#>  ¦   ¦   ¦--8) z2 <= 0.51648      
#>  ¦   ¦   °--9) z2 > 0.51648       
#>  ¦   °--5) z1 > 0.14286           
#>  ¦       ¦--10) z1 <= 0.17582     
#>  ¦       °--11) z1 > 0.17582      
#>  ¦           ¦--22) z1 <= 0.24176*
#>  ¦           °--23) z1 > 0.24176* 
#>  °--3) z1 > 0.27473               
#>      ¦--6) z2 <= 0.46154          
#>      ¦   ¦--12) z2 <= 0.19780     
#>      ¦   °--13) z2 > 0.19780      
#>      ¦       ¦--26) z2 <= 0.25275*
#>      ¦       °--27) z2 > 0.25275* 
#>      °--7) z2 > 0.46154           
#>          ¦--14) z2 <= 0.97802     
#>          ¦   ¦--28) z1 <= 0.74725*
#>          ¦   °--29) z1 > 0.74725* 
#>          °--15) z2 > 0.97802*     
#> 

print(fit, 2)

#>  ROC-guided survival forest: tree # 2 
#> 
#>  node), split
#>    * denotes terminal node
#>   
#> Root                              
#>  ¦--2) z1 <= 0.33000              
#>  ¦   ¦--4) z1 <= 0.11000          
#>  ¦   ¦   ¦--8) z1 <= 0.090000     
#>  ¦   ¦   °--9) z1 > 0.090000*     
#>  ¦   °--5) z1 > 0.11000           
#>  ¦       ¦--10) z1 <= 0.18000*    
#>  ¦       °--11) z1 > 0.18000      
#>  °--3) z1 > 0.33000               
#>      ¦--6) z1 <= 0.83000          
#>      ¦   ¦--12) z1 <= 0.56000     
#>      ¦   °--13) z1 > 0.56000      
#>      ¦       ¦--26) z2 <= 0.59000*
#>      ¦       °--27) z2 > 0.59000  
#>      °--7) z1 > 0.83000           
#>          ¦--14) z2 <= 0.31000     
#>          ¦   ¦--28) z1 <= 0.98000*
#>          ¦   °--29) z1 > 0.98000* 
#>          °--15) z2 > 0.31000      
#>              ¦--30) z2 <= 0.94000*
#>              °--31) z2 > 0.94000* 
#>