Function to generate simulated data used in the manuscript.

This function is used to generate simulated data in the manuscript. The underlying model is assumed to be $$\lambda(t, Z) = \lambda_0(t) e^{(\beta_1 * Z_1 + \beta_2 * Z_2)},$$ where \(\lambda_0(t)\) is the baseline hazard function, (\(b_1, b_2\)) is the regression coefficient, and \(Z_1\) and \(Z_2\) are the covariates. When time varying covariate is involved (scenario 2), only \(Z_1\) is assumed to dependent on time.

This function is used to generate the true cumulative hazard function for ONE ID used in the simulation.

This function is used to generate the true survival function for ONE ID used in the simulation.

This function is used to generate testing sets for each scenario.

simu(n, cen, scenario, summary = FALSE)

trueHaz(dat)

trueSurv(dat)

simuTest(dat)

Arguments

n
is the number of subject

cen
is the censoring percentage; right now it can be either 0%, 25%, or 50%.

scenario
can be numeric or character string. This indicates the simulation scenario noted in the manuscript. See Details for all options.

summary
a logical value indicating whether a brief data summary will be printed.

dat
is a data.frame prepared by simu.

Value

simu returns a data.frame in the class of "roc.simu". This is needed for trueHaz and trueSurv. The returned data.frame consists of columns:

Y

is the observed follow-up time.

death

is the death indicator; death = 0 if censored.

z1

is the time-independent covariate.

z2

is the secondary covariate. It is time-independent in Scenario 1 and time-dependent in Scenario 2 and 3.

Details

The following scenarios are included.

Scenario 1

assumes time independent covariates with \(\Lambda(0) = 2t\):

Scenario 2

assumes time dependent covariate (\(Z_1\)) with \(\Lambda(0) = 2t\). The survival times are generated from the hazard \(\lambda(t, Z(t)) = \lambda_0(t)e^{Z_1(t) + Z_2}\), where \(Z_1(t)\) is the time-dependent covariate.