Function to generate simulated data used in the manuscript.

This function is used to generate simulated data under various settings. Let \(Z\) be a \(p\)-dimensional vector of possible time-dependent covariates and \(\beta\) be the vector of regression coefficient. The survival times (\(T\)) are generated from the hazard function sepecified as follow:

Scenario 1.1

Proportional hazards model: $$\lambda(t|Z) = \lambda_0(t) e^{-0.5 Z_1 + 0.5 Z_2 - 0.5 Z_3 ... + 0.5 Z_{10}},$$

where \lambda_0(t) = 2t.
Scenario 1.2

Proportional hazards model with noise variable: $$\lambda(t|Z) = \lambda_0(t) e^{2Z_1 + 2Z_2 + 0Z_3 + ... + 0Z_{10}},$$

where \lambda_0(t) = 2t.
Scenario 1.3

Proportional hazards model with nonlinear covariate effects: $$\lambda(t|Z) = \lambda_0(t) e^{[2\sin(2\pi Z_1) + 2|Z_2 - 0.5|]},$$

where \lambda_0(t) = 2t.
Scenario 1.4

Accelerated failure time model: $$\log(T) = -2 + 2Z_1 + 2Z_2 + \epsilon,$$ where \(\epsilon\) follows \(N(0, 0.5^2).\)

Scenario 1.5

Generalized gamma family: $$T = e^{\sigma\omega},$$ where \(\omega = \log(Q^2 g) / Q\), \(g\) follows Gamma(\(Q^{-2}, 1\)), \(\sigma = 2Z_1, Q = 2Z_2.\)

Scenario 2.1

Dichotomous time dependent covariate with at most one change in value: $$\lambda(t|Z(t)) = \lambda_0(t)e^{2Z_1(t) + 2Z_2},$$ where \(Z_1(t)\) is the time-dependent covariate: \(Z_1(t) = \theta I(t \ge U_0) + (1 - \theta) I(t < U_0)\), ,\(\theta\) is a Bernoulli variable with equal probability, and \(U_0\) follows a uniform distribution over \([0, 1]\).

Scenario 2.2

Dichotomous time dependent covariate with multiple changes: $$\lambda(t|Z(t)) = e^{2Z_1(t) + 2Z_2},$$ where \(Z_1(t) = \theta[I(U_1\le t < U_2) + I(U_3 \le t)] + (1 - \theta)[I(t < U_1) + I(U_2\le t < U_3)]\), \(\theta\) is a Bernoulli variable with equal probability, and \(U_1\le U_2\le U_3\) are the first three terms of a stationary Poisson process with rate 10.

Scenario 2.3

Proportional hazard model with a continuous time dependent covariate: $$\lambda(t|Z(t)) = 0.1 e^{Z_1(t) + Z_2},$$ where \(Z_1(t) = kt + b\), \(k\) and \(b\) are independent uniform random variables over \([1, 2]\).

Scenario 2.4

Non-proportional hazards model with a continuous time dependent covariate: $$\lambda(t|Z(t)) = 0.1 \cdot[1 + \sin\{Z_1(t) + Z_2\}],$$ where \(Z_1(t) = kt + b\), \(k\) and \(b\) follow independent uniform distributions over \([1, 2]\).

Scenario 2.5

Non-proportional hazards model with a nonlinear time dependent covariate: $$\lambda(t|Z(t)) = 0.1 \cdot[1 + \sin\{Z_1(t) + Z_2\}],$$ where \(Z_1(t) = 2kt\cdot \{I(t > 5) - 1\} + b\), \(k\) and \(b\) follow independent uniform distributions over \([1, 2]\).

The censoring times are generated from an independent uniform distribution over [0, c], where c was tuned to yield censoring percentages of 25% and 50%.

simu(n, cen, scenario, summary = FALSE)

trueHaz(dat)

trueSurv(dat)

Arguments

n
an integer value indicating the number of subjects.

cen
is a numeric value indicating the censoring percentage; three levels, 0%, 25%, 50%, are allowed.

scenario
can be either a numeric value or a character string. This indicates the simulation scenario noted above.

summary
a logical value indicating whether a brief data summary will be printed.

dat
is a data.frame prepared by simu.

Value

simu returns a data.frame. The returned data.frame consists of columns:

id

is the subject id.

Y

is the observed follow-up time.

death

is the death indicator; death = 0 if censored.

z1--z10

is the possible time-independent covariate.

k, b, U

are the latent variables used to generate $Z_1(t)$ in Scenario 2.1 -- 2.5.

The returned data.frame can be supply to trueHaz and trueSurv to generate the true cumulative hazard function and the survival function, respectively.

Examples

set.seed(1)
simu(10, 0.25, 1.2, TRUE)

#> Summary results:
#> Number of subjects: 10
#> Number of subjects experienced death: 9
#> Number of covariates: 10
#> Time independent covaraites: z1 z2 z3 z4 z5 z6 z7 z8 z9 z10
#> Number of unique observation times: 10
#> Median survival time: 0.2983867
#> 

#> # A tibble: 55 x 13
#>       id   Time death    z1    z2    z3    z4    z5    z6    z7    z8    z9
#>    <int>  <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1     1 0.0822     0 0.266 0.206 0.935 0.482 0.821 0.478 0.913 0.339 0.435
#>  2     1 0.123      0 0.266 0.206 0.935 0.482 0.821 0.478 0.913 0.339 0.435
#>  3     1 0.238      0 0.266 0.206 0.935 0.482 0.821 0.478 0.913 0.339 0.435
#>  4     1 0.239      0 0.266 0.206 0.935 0.482 0.821 0.478 0.913 0.339 0.435
#>  5     1 0.250      0 0.266 0.206 0.935 0.482 0.821 0.478 0.913 0.339 0.435
#>  6     1 0.347      1 0.266 0.206 0.935 0.482 0.821 0.478 0.913 0.339 0.435
#>  7     2 0.0822     0 0.372 0.177 0.212 0.600 0.647 0.861 0.294 0.839 0.713
#>  8     2 0.123      0 0.372 0.177 0.212 0.600 0.647 0.861 0.294 0.839 0.713
#>  9     2 0.238      0 0.372 0.177 0.212 0.600 0.647 0.861 0.294 0.839 0.713
#> 10     2 0.239      0 0.372 0.177 0.212 0.600 0.647 0.861 0.294 0.839 0.713
#> # ... with 45 more rows, and 1 more variable: z10 <dbl>

set.seed(1)
simu(10, 0.50, 2.2, TRUE)

#> Summary results:
#> Number of subjects: 10
#> Number of subjects experienced death: 7
#> Number of covariates: 2
#> Time independent covaraites: z1.
#> Time dependent covaraites: z2.
#> Number of unique observation times: 10
#> Median survival time: 0.04646691
#> 

#> # A tibble: 55 x 9
#>       id    Time death    z1    z2     e    u1    u2    u3
#>    <int>   <dbl> <dbl> <dbl> <dbl> <int> <dbl> <dbl> <dbl>
#>  1     1 0.00183     0     0 0.206     1 0.104 0.109 0.212
#>  2     1 0.00536     0     0 0.206     1 0.104 0.109 0.212
#>  3     1 0.00604     0     0 0.206     1 0.104 0.109 0.212
#>  4     1 0.00771     0     0 0.206     1 0.104 0.109 0.212
#>  5     1 0.0335      0     0 0.206     1 0.104 0.109 0.212
#>  6     1 0.0465      0     0 0.206     1 0.104 0.109 0.212
#>  7     1 0.0675      0     0 0.206     1 0.104 0.109 0.212
#>  8     1 0.0757      0     0 0.206     1 0.104 0.109 0.212
#>  9     1 0.0940      0     0 0.206     1 0.104 0.109 0.212
#> 10     1 0.264       0     1 0.206     1 0.104 0.109 0.212
#> # ... with 45 more rows